init SimulationParams
